requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
audiomentations==0.42.0
audioop-lts==0.2.1
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
numba==0.61.2
numpy==2.2.6
numpy-minmax==0.5.0
numpy-rms==0.6.0
openai-whisper==20250625
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-stretch==0.3.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sentry-sdk==2.32.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
xxhash==3.5.0
yarl==1.20.1
